+++ drivers/basic/src/main/java/de/learnlib/driver/api/TestDriver.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.learnlib.drivers.api;
+package de.learnlib.driver.api;
 
 import de.learnlib.api.SUL;
 import de.learnlib.mapper.ExecutableInputSUL;