+++ drivers/basic/src/main/java/de/learnlib/driver/reflect/MethodOutput.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.learnlib.drivers.reflect;
+package de.learnlib.driver.reflect;
 
 /**
  * abstract output of a method.