+++ drivers/basic/src/test/java/de/learnlib/driver/reflect/ObjectTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.learnlib.drivers.reflect;
+package de.learnlib.driver.reflect;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;